Add last page link in paginated tables
When working on workflows that have many instances the most recent one that ususally is of interest can be many pages ahead. People have to click multiple times to get to the last page or do math and manipulate manually the pageNumber in the url parameters neither of which are ideal.